Dog-Breed-Identification-using-Transfer-Learning-main/requirements.txt
numpy==1.23.5
pandas==1.5.3
flask==2.3.3
tensorflow==2.12.0
opencv-python==4.11.0.86
pillow==10.4.0
scikit-learn==1.2.2